+++ node.h
@@ -2,6 +2,6 @@
 #define FUNCTIONS_NODE_H
 
 typedef struct node node;
-struct node* createNode(int value, int type); // creates a new node with the input value
+struct node* createNode(double value, int type); // creates a new node with the input value
 
 #endif // FUNCTIONS_NODE_H
